requirements.txt
requests
python-dotenv
pandas
numpy
scikit-learn
matplotlib
jupyter